:recycle: [Refactor]: features/notification -> features/view-notification 이름 변경
- entities의 notification과 혼동 우려가 있어서 변경
